+++ pkg/analyzer/query.go
@@ -157,6 +157,9 @@ func (query *queryBuilder) selectFields() bool {
 					query.args = append(query.args, query.filter.EventMetaKey[0])
 					q.WriteString(fmt.Sprintf("%s %s,", query.selectField(query.fields[i]), query.fields[i].Name))
 				}
+			} else if query.fields[i] == FieldEventMetaCustomMetric {
+				query.args = append(query.args, query.filter.CustomMetricKey)
+				q.WriteString(fmt.Sprintf("%s %s,", fmt.Sprintf(query.selectField(query.fields[i]), query.filter.CustomMetricType), query.fields[i].Name))
 			} else if query.parent != nil && (query.fields[i] == FieldEntryTitle || query.fields[i] == FieldExitTitle) {
 				q.WriteString(query.selectField(query.fields[i]) + ",")
 			} else {