+++ pkg/analyzer/query_test.go
@@ -174,3 +174,27 @@ func TestQueryPlatformPageView(t *testing.T) {
 	assert.Equal(t, "/foo", args[13])
 	assert.Equal(t, "SELECT toInt64OrDefault((SELECT uniq(t.visitor_id) visitors FROM page_view t JOIN (SELECT visitor_id visitor_id,session_id session_id FROM session t WHERE client_id = ? AND toDate(time, 'UTC') >= toDate(?) AND toDate(time, 'UTC') <= toDate(?) GROUP BY visitor_id,session_id HAVING sum(sign) > 0 ) j ON j.visitor_id = t.visitor_id AND j.session_id = t.session_id WHERE client_id = ? AND toDate(time, 'UTC') >= toDate(?) AND toDate(time, 'UTC') <= toDate(?) AND desktop = 1 AND mobile = 0 AND path = ? )) platform_desktop,toInt64OrDefault((SELECT uniq(t.visitor_id) visitors FROM page_view t JOIN (SELECT visitor_id visitor_id,session_id session_id FROM session t WHERE client_id = ? AND toDate(time, 'UTC') >= toDate(?) AND toDate(time, 'UTC') <= toDate(?) GROUP BY visitor_id,session_id HAVING sum(sign) > 0 ) j ON j.visitor_id = t.visitor_id AND j.session_id = t.session_id WHERE client_id = ? AND toDate(time, 'UTC') >= toDate(?) AND toDate(time, 'UTC') <= toDate(?) AND desktop = 0 AND mobile = 1 AND path = ? )) platform_mobile ", queryStr)
 }
+
+func TestQueryCustomMetric(t *testing.T) {
+	filter := &Filter{
+		ClientID:         42,
+		From:             util.PastDay(7),
+		To:               util.Today(),
+		EventName:        []string{"Event"},
+		CustomMetricType: pkg.CustomMetricTypeFloat,
+		CustomMetricKey:  "Custom Meta Value",
+	}
+	q := queryBuilder{
+		filter: filter,
+		fields: []Field{
+			FieldEventMetaCustomMetric,
+			FieldVisitors,
+		},
+		from: events,
+	}
+	queryStr, args := q.query()
+	assert.Len(t, args, 5)
+	assert.Equal(t, "Custom Meta Value", args[0])
+	assert.Equal(t, "Event", args[4])
+	assert.Equal(t, "SELECT coalesce(toFloat64OrZero(event_meta_values[indexOf(event_meta_keys, ?)])) custom_metric,uniq(t.visitor_id) visitors FROM event t WHERE client_id = ? AND toDate(time, 'UTC') >= toDate(?) AND toDate(time, 'UTC') <= toDate(?) AND event_name = ? ", queryStr)
+}