+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * added support for client hints
 * added Chrome OS and Windows 11 detection
 * added conversion rate to total, by period, and growth
+* added custom metrics for events
 * updated User-Agent blacklist
 * updated referrer blacklist
 * upgraded to Go 1.21